Renamed all the Shopping.Aggregator servies as ProxyServices
