+++ src/ApiGateways/Shopping.Aggregator/Models/BasketModel.cs
@@ -10,11 +10,11 @@ public class BasketModel
 
     public decimal? TotalPrice { get; set; }
 
-    public string? Category { get; set; }
+    // public string? Category { get; set; }
 
-    public string? Summary { get; set; }
+    //public string? Summary { get; set; }
 
-    public string? Description { get; set; }
+    //public string? Description { get; set; }
 
-    public string? ImageFile { get; set; }
+    // public string? ImageFile { get; set; }
 }
