+++ src/ApiGateways/Shopping.Aggregator/ProxyServices/BasketService.cs
@@ -1,8 +1,7 @@
-﻿using System;
-using Shopping.Aggregator.Extensions;
+﻿using Shopping.Aggregator.Extensions;
 using Shopping.Aggregator.Models;
 
-namespace Shopping.Aggregator.Services;
+namespace Shopping.Aggregator.ProxyServices;
 
 public class BasketService : IBasketService
 {
