+++ src/ApiGateways/Shopping.Aggregator/ProxyServices/IBasketService.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Shopping.Aggregator.Models;
 
-namespace Shopping.Aggregator.Services;
+namespace Shopping.Aggregator.ProxyServices;
 
 public interface IBasketService
 {
