+++ src/Services/Catalog/Catalog.Api/Controllers/v1/CatalogController.cs
@@ -29,7 +29,7 @@ namespace Catalog.Api.Controllers.v1
         [HttpGet("{id:length(24)}")]
         [ProducesResponseType((int)HttpStatusCode.NotFound)]
         [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
-        public async Task<ActionResult<Product>> GetProductsById(string id)
+        public async Task<ActionResult<Product>> GetProductById(string id)
         {
             var product = await this._repository.GetProduct(id);
